#10: refactoring
