--- compoly-bot/src/main/kotlin/com/vitekkor/compolybot/model/telegram/MediaGroup.kt
@@ -1,5 +1,0 @@
-package com.vitekkor.compolybot.model.telegram
-
-data class MediaGroup(
-    val media: List<InputMedia>
-)
