--- compoly-bot/src/main/kotlin/com/vitekkor/compolybot/scenario/extension/Beans.kt
@@ -1,7 +1,0 @@
-package com.vitekkor.compolybot.scenario.extension
-
-import com.vitekkor.compolybot.scenario.command.BaseCommand
-import com.vitekkor.compolybot.service.TelegramApiService
-import com.vitekkor.compolybot.service.bean
-
-val telegramApiService by bean<TelegramApiService>()
